gcc: Additional warning cleanup
This commit addresses some issues encountered when adding the
`-Wconversion` and `-Wno-sign-conversion` flags for stricter type
and sign conversion checks.

It also removes the `ZSL_MTX_UNARY_OP_LOGICAL_NEGATION` unary operand
in matrices.s since it is of marginal use in the real world, and causes
a cast issue from float to bool.

Signed-off-by: Kevin Townsend <kevin@ktownsend.com>