config: Removed CONFIG_ZSL_MATRIX_QRD_USE_SCRATCH
Removed the `CONFIG_ZSL_MATRIX_QRD_USE_SCRATCH` config
flag, and allocate everything on the stack for now. Future
work will need to be done to reign in stack memory use for
the iterative eigenvalue and eigenvector functions used
by pinv and svd.

Increased the stack size for the pinv and svd samples to
80 KB so that they can be run in qemu.

Signed-off-by: Kevin Townsend <kevin@ktownsend.com>