orientation: ahrs: fix an error in zsl_att_from_accelmag
The angles for calculating the tilt compensation were swapped.

When compensating the x-axis like in den (mag_unit.data[0]) the roll
has no effect as this is the angle around which the object is turned.
The same for the y-axis and pitch.

This PR fixes the assignment of angles to axis and updates the affected unit tests. The update in zsl_att_from_accelmag has been tested in an implementation.

Signed-off-by: Stefan Schmidt <kontakt@stefanschmidt-embedded.de>